added suspend handling

+++ Common.py
@@ -76,7 +76,7 @@ def blitMultilineText(surface, text, pos, font, color=pygame.Color('black')):
 def quick_sort(items):
         """ Implementation of quick sort """
         if len(items) > 1:
-                pivot_index = len(items) / 2
+                pivot_index = int(len(items) / 2)
                 smaller_items = []
                 larger_items = []
